fixed bug where 'long' was 8 bytes. changed it to 4 bytes and added keyword(s) 'long long' which is 8 bytes

+++ src/headers/stddef.h
@@ -1,8 +1,8 @@
 #ifndef _VBRCC_STDDEF_H
 #define _VBRCC_STDDEF_H
 
-typedef long size_t;
-typedef long ptrdiff_t;
+typedef long long size_t;
+typedef long long ptrdiff_t;
 
 #define NULL ((void *)0)
